requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2

qdrant-client==1.7.3
sentence-transformers==2.7.0
google-generativeai==0.5.4